Remove AMD_performance_monitor + INTEL_performance_query
There are a few reasons for taking a clean slate approach here:

Neither of these extensions is a neat subset of the other which makes
having their implementations be intertwined a bit more cumbersome than
it needs to be. Not all of the semantics types exposed by these
extensions are currently supported in Mesa but doing so either means
adding some way of white listing counters for each extension in the
backend or finding some way of mapping the types between the extensions.

There is only one backend which isn't really usable in its current state
because it reports incorrect counter data (observability counters).  It
seems worthwhile simplifying things and focusing on getting one
extension working well.

Other considerations here include:
- Wanting to give the backend full freedom over what data structures it
  uses to track queries and counters, avoiding the need for a common
  denominator structure that mesa/main can poke at. This is because we
  know counter descriptions are going to get more complex as we look at
  adding more interesting OA counters to the brw backend.
- wanting to re-jig the entry points into the backend to be more in line
  with how query objects work, including a vfunc for explicitly waiting
  for a query instead assuming glFinish() is enough.
- wanting to get rid of the need to track active counters/groups on a
  per query object basis which is an added complexity that's not
  required for the INTEL_performance_query extension.