i965: Expose OA counters via INTEL_performance_query
This adds support for exposing basic Observation Architecture
performance counters on Haswell.

This support is based on the i915_oa perf kernel interface which is used
to configure the OA unit, allowing Mesa to emit MI_REPORT_PERF_COUNT
commands around queries to collect counter snapshots.

To take into account the small chance that some of the 32bit counters
could wrap around for long queries (~50 milliseconds for a GT3 Haswell @
1.1GHz) the implementation also collects periodic metrics via perf.

It seems worth noting that the structures chosen to describe the
counters were arrived at after a few iterations of creating other
limited intel_gpu_tools programs that similarly needed to describe,
normalise and display these counters and many of the utility functions
should be reusable as more sets of counters are exposed via the OA unit
as well as for future hardware generations.

It also seems worth noting that in the future we might want to aim to
factor out much of the code for describing and normalising OA counters
into a common library such as libdrm since we are already having
difficulties copying very similar code between multiple tools dealing
with these same counters; such as for looking at the OA counters for
system wide analysis. Until we've got a bit more experience dealing with
the counters though it seems worth deferring this separation and having
to design a suitable api for these different use cases.